# server/api/swagger.yaml
basePath: /
definitions:
  Agent:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      isOnline:
        type: boolean
      lastSeenAt:
        type: string
      name:
        type: string
    type: object
  CreateAgentBody:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  CreateAgentResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      enrollmentKey:
        type: string
      id:
        type: string
      isOnline:
        type: boolean
      lastSeenAt:
        type: string
      name:
        type: string
    type: object
  gin.H:
    additionalProperties: {}
    type: object
host: localhost:8091
info:
  contact: {}
  title: Monitoring Server API
  version: "1.0"
paths:
  /agents:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Agent'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get all agents
    post:
      consumes:
      - application/json
      parameters:
      - description: Create agent body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateAgentBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CreateAgentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create new agent
swagger: "2.0"